Move hook to hook file
