+++ src/custom/state/application/atom.ts
@@ -1,6 +1,5 @@
-import { atom, useAtom } from 'jotai'
+import { atom } from 'jotai'
 import { atomWithStorage } from 'jotai/utils'
-import { useCallback } from 'react'
 
 type FollowPendingTxPopup = {
   showPopup: boolean
@@ -8,7 +7,7 @@ type FollowPendingTxPopup = {
 }
 
 /**
- * Base atom that store the modal that allow to follow a pending tx
+ * Base atom that store the popup state that indicate how to follow a pending tx
  */
 export const followPendingTxPopupAtom = atomWithStorage<FollowPendingTxPopup>('followPendingTxPopup', {
   hideFollowTxPopup: false,
@@ -26,29 +25,3 @@ export const handleHidePopupPermanentlyAtom = atom(null, (_get, set, update: boo
 export const showFollowTxPopupAtom = atom(
   (get) => get(followPendingTxPopupAtom).showPopup === true && !get(followPendingTxPopupAtom).hideFollowTxPopup
 )
-
-export const useFollowPendingTxPopup = () => {
-  const [, _setFollowPendingPopup] = useAtom(handleFollowPendingTxPopupAtom)
-  const [, _setHidePopup] = useAtom(handleHidePopupPermanentlyAtom)
-  const [_showFollowPendingTxPopup] = useAtom(showFollowTxPopupAtom)
-
-  const setShowFollowPendingTxPopup = useCallback(
-    (showPopup: boolean) => {
-      _setFollowPendingPopup(showPopup)
-    },
-    [_setFollowPendingPopup]
-  )
-
-  const setHidePendingTxPopupPermanently = useCallback(
-    (hidePopup: boolean) => {
-      _setHidePopup(hidePopup)
-    },
-    [_setHidePopup]
-  )
-
-  return {
-    showFollowPendingTxPopup: _showFollowPendingTxPopup,
-    setShowFollowPendingTxPopup,
-    setHidePendingTxPopupPermanently,
-  }
-}
